requirements.txt
streamlit
langchain_core
groq
langchain_groq
python-dotenv
langchain_community
pypdf
sentence-transformers
faiss-cpu